Fix repo configuration
The wrong parameter name was used for the repo configuration.
It should have been rpm_setup_config not rpm_setup. Additionally the
iptables rules did not allow for the containers to fetch the repos from
the CI host. This change lets docker manage the docker iptables rules
and adds a rule to allow port 80 from the docker container ip range.

Depends-On: https://review.openstack.org/#/c/634064/
Change-Id: I037465f7bbc7b6016373385f07c7c944a09299c9
Closes-Bug: #1813527