Use override-checkout for stable branch jobs
Now that we have tripleo-upgrade as a required-project [1] it is
always cloned by zuul, so toci old mechanism [2] to detect what to clone
is obsolete.

Instead of that we can use zuul override-checkout [3] that checkouts a
specified branch of the cloned projects, which is exactly the case of
branchful jobs (-rocky, -queens, -pike, etc...)

NOTE: We have also to fix this at RDO branchful jobs.

[1] https://github.com/openstack-infra/tripleo-ci/blob/master/zuul.d/base.yaml#L26-L37a
[2] https://github.com/openstack-infra/tripleo-ci/blob/master/roles/run-test/templates/toci_gate_test.sh.j2#L155-L161
[3] https://www.zuul-ci.org/docs/zuul/user/config.html?highlight=override%20checkout

Depends-On: https://review.openstack.org/#/c/634724

Related-Bug: #1814531

Change-Id: I2aff01f371d41b48b0f364ed91adf691a74bfc68