Parameterize build-containers job to allow push to registry
As tracked by ci team in [1] the current periodic container
build job is very different to the upstream job
tripleo-build-containers-centos-7. In order to re-use this
upstream job as part of the periodic, we need to parameterize
it so we can push to the rdo registry and containers have the
right tags (when periodic container build & push).

This review leaves the upstream job unchanged with parameters
defaulting to their current value. The periodic job defined
in [2] inherits from this job and will handle secrets/login
and passing the right values for registry/tag.

[1] https://tree.taiga.io/project/tripleo-ci-board/us/652
[2] https://review.rdoproject.org/r/#/c/18729
Change-Id: Idf1f5a696a5163bd54cb000aed09daf00d51400b