Make build-containers job definition reusable
Allows us to define build-containers-fedora-28 job on rdo that would
inherit the upstream definition.

Base jobs that do have file patterns define should have a
`-base-with-files` suffix instead of just `-base`, so we can avoid
accidental use in periodic jobs.

Adds missing patterns on triple-repos which is used by build-containers
in order to avoid making changes to it that would break these jobs.

Change-Id: Ibd0512f1e9bc9c4acb8c0256d1f161bd6d09fe73
Needed-By: https://review.rdoproject.org/r/#/c/18913/
Related-To: https://storyboard.openstack.org/#!/story/2005040