Add override_repos into tripleo-ci tripleo-repos role
The tripleo-repos role in tripleo-ci defaults to current-tripleo
for master and current for everything else. There is no support
for the tripleo-ci-testing repo.

This adds 2 params: i) override_repos as a param
to tripleo-ci tripleo-repos and ii) buildcontainers_override_repos
as a param in the tripleo-buildcontainers job. ii) is passed from
periodic version of buildcontainers in [1] and is then used to
set the value of override_repos, passed into ci tripleo-repos.

The depends-on adds tripleo-ci-testing as an option for
tripleo-repos itself as it is currently unsupported.
Tracked by tripleo ci squad with [2]. See test review at [3] which
runs the periodic container build in check with this code.

Also adds /etc/yum.repos.d/ into collected buildcontainer logs.

[1] https://review.rdoproject.org/r/18975
[2] https://tree.taiga.io/project/tripleo-ci-board/task/773
[3] https://review.rdoproject.org/r/19000
Depends-On: https://review.openstack.org/638651 # tripleo-repos
Change-Id: I37bc90881ef19ae4aeace65750da4dd93da5f4e2