Create two loop devices for bluestore and filestore
The ceph-loop-device role was changed [1] to create an
LVM structure on the loop device but only for versions
which can use that structure for bluestore. However,
the zuul.branch variable doesn't work as expected to
differentiate versions so remove the code that uses it.

In place of switching on the version create two loop
devices, then legacy jobs which don't support LVM can
use the loop device without LVM. Call the legacy loop
device /dev/loop3 so older jobs get the expected name.

[1] https://review.openstack.org/#/c/637196

Change-Id: I778bf3d78ccf492e852a2c932ccd5e8266bcfe71
Related-Bug: #1817688