Remove retag tripleo-ci-testing for periodic containers build job
As described in the bug the retag added at [1] isn't working.

However we are correctly pushing with the tripleo-ci-testing delorean
repo hash and that should be sufficient for the periodic jobs that
consume containers built by this job.

This removes the retag logic altogether and also removes the
push_tag variable. A related review in [2] also removes it in the
periodic jobs that were using this variable (no one else does/was)

Related-Bug: 1819583
[1] https://review.openstack.org/641348
[2] https://review.rdoproject.org/r/19484
Change-Id: I42f8e05bf1e5be230acdb7be16b056c8aff117f2