Adds conditional skip excludes in containers build for rocky
The excludes list is needed for f28 containers build. For the
rocky periodic we need to skip it as there is no such parameter
in the container build there.

Instead uses sed to remove excludes from containers list.
Once we cherrypick that [1] we can remove sed and use --excludes.

Test runs in [2]

[1] https://review.openstack.org/#/q/topic:cherrypick-excludes-containers-build
[2] https://review.rdoproject.org/r/#/c/19533/
Co-Authored-By: Sagi Shnaidman <sshnaidm@redhat.com>
Change-Id: I668a8557165972324e11707c2f3f32199658b6b9