Rocky containers branchfull job
As tripleo-ci repo is branchless and has only master, we need
a rocky containers job that will run on this repo and will check
that we don't break anything for rocky containers building.
Breaking this job will affect periodic promotion because we started
to use this job for containers building in promotions.

Depends-On: https://review.openstack.org/#/c/648358
Change-Id: I8a1c69a9ab1de6992eb52b76862cacacaf265896