Include vars of tempest_services and concurrency in fs override
Since we are moving standalone scenario jobs to os_tempest for
that we need to put the needed os_tempest vars in featureset
override so that it can be used in job definition.
Adding tempest_plugin_distro_packages for tempest plugins
installation and tempest_run_concurrency for setting
tempest workers.

https://tree.taiga.io/project/tripleo-ci-board/task/906

Depends-On: https://review.openstack.org/#/c/649577/

Change-Id: I06cdae5cbdc0073323b069bf4b03a6773369727b