Allow to override "validate_services" param in featureset
This will allow to enable that new validation only for the non-voting
jobs.

Doing it at that higher level is better than editing the featuresets,
since one set can be used in both voting and non-voting jobs.

This is a temporary activation, since we need to gather non-blocking
information about that new validation.

Change-Id: I42e2e5282e1406385173ad3e5a79a76c2bb02810