Do not run reset_connection meta with buildah
reset_connection is not needed when running container
build with buildah.

reset_conneciton doesn't work[1] with when condition and has
different behavior with different ansible versions:-
- With ansible-2.5.15:- Fatal error
- With ansible-2.7:- Warning

Since the ansible version in RDO zuul is still 2.5 we
need to workaround by moving the meta task to different
play and include it in the main playbook.

[1] https://github.com/ansible/ansible/issues/46275

Change-Id: Id6bbf0da1c504dcf1bb09c9435c92e0f18c7c06b