Introduce arch aware tag for container images
We will soon be pushing ppc64le container images to
registry.rdoproject.org and need to tag images uniquely from
the x86_64 versions. This patch adds a suffix of the format
'_{{ansible_architecture}}' to the tag of the images. To ensure
backwards compatibility, x86_64 will also use the current tag
of format {{ version_hash }}. The retagging ansible scripts
are modified from an older patchset [0]. Upon promotion, the
suffix should be removed from the tag, a manifest should be
created, and the images should be uploaded to dockerhub.

[0] https://review.opendev.org/#/c/660665/

Depends-On: https://review.opendev.org/678907

Co-Authored-By: Rafael Folco <rfolco@redhat.com>

Change-Id: I537f577c8946ffeeed989a7c686ff85fef96a3f7