Add conditional periodic-ness to tripleo jobs
Sometimes we need to run a periodic job with repositories like
it's in regular check jobs, for example periodic jobs in tripleo-ci
which are forced currently to run with promotion repos:
"force_non_periodic" variable in job shall force job to run as
"check" and use regular repositories as any check job.

Sometimes we need to run a check job with promotion repositories
to debug it or to try it, instead of depending on some patch
let's use a job variable for that:
"force_periodic" will force job to run as promotion job, using the
same repositories as any promotion job uses.

Change-Id: Ib305207dcd2b4626a7d7d340bc383dd176b11633