Vote and gate on containers-multinode-branchful jobs
Currently containers multinode jobs (stein and rocky) are
non-voting and only in check. There is no reason for not
voting and gating on containers multinode branchful jobs.

http://zuul.openstack.org/builds?job_name=tripleo-ci-centos-7-containers-multinode-stein&job_name=tripleo-ci-centos-7-containers-multinode-rocky

These jobs belong to (and are controlled by) the template
'tripleo-multinode-branchful':

 - tripleo-ci-centos-7-containers-multinode-stein
 - tripleo-ci-centos-7-containers-multinode-rocky
 - tripleo-ci-centos-7-containers-multinode-queens

NOTE: These jobs should not be used outside of ci realm,
that is, the template should be used in ci repos only
(tq, tqe, and tripleo-ci).

Change-Id: Ie809f8bbb4e62f5d3b37efc026b8efeb74d9dd92