Fixed containers-build role to compare built containers
Since podman is not installed by default in containers-build job
so the Built var is returning zero and further giving
[: -l: binary operator expected error.

In order to fix the above issue, It adds task
to set proper container_cli and reuse buildah and docker
and compare the build containers status using returncode.

Use sudo while cating the files as buildah is runned via
sudo and it will fix permission issue.

Change-Id: I61cd114ec7150d2c6156c5d8a2fe105fdd458348
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>