Fix dumping variables when no featureset is defined
In some jobs we don't define featureset, like building images
and containers. In this case we need to set up variables with
option that featureset doesn't exist.
Add checks for featureset defined into common role.
Also add tripleo roles that are used in building containers job
for run condition for these jobs. Add playbooks to be gated as
well.

Change-Id: I24364bda1d47c21259333bc936101e203a1689ca