Disable voting on container-updates until 678695 is merged
We have a chicken and egg issue where we both need
https://review.opendev.org/#/c/678695/ and
https://review.opendev.org/#/c/679299/ to land so we can ahve the
container-updates job running on Podman and not Docker.

The reason is because we can't run Podman when Pacemaker is enabled on
CentOS7, so 679299 would never pass. We are going to break the CI job
when 679299 merges but 679299 will fix it by disabling Pacemaker.

This patch will be reverted soon after 678695 is merged.

Change-Id: Ie7e1a78238fcea22032bda1162006582dc8c1fe3