Add test_snapshot_pattern to scenario001 tempest whitelist
With bug #1840138 we tried to disable use of multiple_locations in
Glance to avoid security risks but it is necessary to keep that
feature enabled when Glance is backed by RBD or Nova won't be able
to publish new snapshots location, see bug #1842938.

This change whitelists an image snapshot test from tempest in
scenario001 to have some coverage for the above situation.

Change-Id: I5e2c19df59581e1ead84a31107e59f6536b490ff
Co-Authored-By: Chandan Kumar <chkumar@redhat.com>