Add component_ci_containers in build-containers for component ci
As part of [1] we need to override the list of containers built
by the build-containers role. This adds a component_ci_containers
parameter that is passed by the job at [2]. This sets the
containers file added here for the compute component (it was
initially at [3] but moved here after initial discussion).

[1] https://tree.taiga.io/project/tripleo-ci-board/task/1378
[2] https://review.rdoproject.org/r/23655
[3] https://review.opendev.org/692840

Change-Id: I60a8cc66c439cbae7b8697a8279996ff82cd4375