Add log collection entry for new ansible file
Durring the container image prepare process the ansible log file will now
be written to a separate file. This change ensures that we collect the file
should it exist. By default the file will be store in the same path as the
created log handler, however in the event that the log handler is only
stream processing the ansible log file will be writtent to /tmp. To ensure
we're collecting the logs correctly, both the /tmp and expected /var/log,
path's have been added to log file array.

Depends-On: Ib208e53619901421d785ca0883f65954f02e59be
Change-Id: Iacc2a015f25266b29feff68bc7ed81b27f9d6aec
Signed-off-by: Kevin Carter <kecarter@redhat.com>