Add option to build containers from package
This patch adds an option to build containers using
tripleoclient and kolla installed via package not
from source keeping default installation via source.

The initial usage of this feature would be to
build containers from cloudsig RDO repos. The idea
is to use same source(packages) for building containers and
content of containers.

Change-Id: Id7cde12a687a1e2a55717ba6fbc0bb78c80c51f6