container-build: use the new --work-dir option
To gather all the needed logs produced by Kolla and Buildah, we
implemented a --work-dir option to generate configs and logs in the same
directory. This patch aims to use it only on master for now. Logs will
be collected the same since we mv /tmp/kolla-* to the artifacts.

Depends-On: https://review.opendev.org/#/c/709039/
Change-Id: I6f5ff7785caf61f5a7ce829099b9258b22bf343a