Make scenario010 voting
Octavia jobs should be voting, we are merging a bunch of patches for
Octavia and they should be properly tested. If something is broken in
the job, it should be fixed; and not ignored by a non voting job.

This patch switches voting to True for
tripleo-ci-centos-8-scenario010-standalone.

Note about tripleo-ci-centos-7-scenario010-ovn-provider-standalone:
the job is being converted to run tempest and is still WIP so it'll vote
later.

Change-Id: I85fdc2fae66b95a194fd3a5da851f702c9dcec1e