Modify build-containers role to allow for internal builds
This modifies the build-containers role to allow us to
build d/stream rhel8 containers. Mainly adds the parameterization
and conditionals to skip/enable parts of the role and then we have
any d/stream specific overrides passed from the job definition.

This also adds a new tripleo-build-containers-base-noreposetup-rhel8
with no required-projects since we need the d/stream equivalents
which we clone in our d/stream job definition as required-projects.

Tracked by tripleo-ci squad in [1].

Co-Authored-By: Marios Andreou <marios@redhat.com>
[1] https://tree.taiga.io/project/tripleo-ci-board/task/1596
Change-Id: I41252222e35565f5c3006eeb38f4dcb82fa980dd