Run virt customize on overcloud image for post sanity
After building images, sometimes few of the files from the installed
packages got missing and there is no way to verify that. Collecting
a list of installed RPMs and missing files will give an idea what
is wrong with the image.

It can enabled or disabled by toggling image_sanity var.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I219454df3ec7a86a3fd033bf384a1f824344646e