Add conditional job fail when logs collection failed
In usual jobs we don't fail jobs if logs collection failed, only
print the message. And that's totally fine, because job should
test TripleO and not logs collection.
But for repos like collect-logs role the logs collection
functionality is critical and when job pass in CI it's possible
that logs collection failed in the patch. Because logs are core
functionality in these repos we should fail the job if logs
collection failed on patch.
For that add variable fail_logs_collection that is false by
default and will be true only for jobs running on logs collection
repository.
In order not to reset whole gate and fail other patches, don't
fail jobs in gate because of logs collection.

Change-Id: I3a946d170c0cb90476e8b23ef0d619fd8d7fcae6