Port sc10 job to os_tempest
https://review.opendev.org/#/c/726752/ ports the fs062 to
os_tempest and c8 based job runs from ussuri onwards so
it is safe to port validate-tempest job var to os_tempest and
remove tempest_worker (validate-tempest) var.

Above review also overrides the sc10-ovn job so the job has
started passing with the tests defined in fs062 as current
tests in jobs are broken so keeping the job nv.

It ports:
* tempest_whitelist -> tempest_test_whitelist
* tempest_extra_config ->  tempest_tempestconf_profile_overrides

tempest_tempestconf_profile_overrides adds as a featureset
overrides as it will be used in zuul jobs.

Note: sc10-ovn job is still broken with failing tempest tests.

Change-Id: I6be42c07b196829b771637d4a2c2f0cb3ee4456c
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>