Adds new tripleo-ci-buildimage-jobs template for branchless repos
This adds a tripleo-ci-buildimage-jobs template for the buildimage
jobs definitions to be used in the branchless repos similar to the
tripleo-ci-build-containers template. Also adds missing branches:
master for -train -ussuri jobs and makes the buildimage hardened
jobs voting. Also marks train jobs as non-voting due to #1881732

This patch was originally posted after discussion in the parent review
at [1]. The Depends-On is for failing build-containers-centos-8-ussuri

[1] https://review.opendev.org/#/c/722790/

Partial-Bug: #1881732
Depends-On: https://review.opendev.org/733790
Change-Id: I85a1c0a102613a88c31cd62554e5cfcc79423b17