Restructure exclude containers based on ci_branch
Currently containers exclude list is maintained based on distro
version and release which seems to be confusing. So we are modifing
the exclude list based on ci_branch so that it would be exclusive and
will be used in promoter config to exclude containers before uploading
images to docker.io.

Use ci_branch for determining release.

Co-authored-by: Bhagyashri Shewale <bshewale@redhat.com>
Change-Id: I66afdf63dc0fd912a9608bbe4528b2574ea58790
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>