Fix container build role to support new container workflow
When push_containers is set to true, it needs to be pushed
to a particular registry with proper namespace, prefix
and tag.

The default prefix is openstack but it needs to be
centos-binary.

Use version_hash to tag the containers and grep successful
containers based on ansible distribution as it defaults points
to kolla base.

It fixes the build.sh.j2 script for the same based on the
following doc[1.]
[1].
https://docs.openstack.org/project-deploy-guide/tripleo-docs/latest/deployment/3rd_party.html#building-new-containers-with-tripleo-container-image-build

It is getting tested here: https://review.rdoproject.org/r/#/c/28194

https://tree.taiga.io/project/tripleo-ci-board/task/1834

Change-Id: I1989f4fcb94a9bc2e8504e8263377aa154efb549
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Co-authored-by: Bhagyashri Shewale <bshewale@redhat.com>