Don't allow to run train on centos7
This change will not allow to run the train on centos7.

Here we are going to migrate upstream stable/train jobs
from CentOS-7 to CentOS-8 [1] is close.
CentOS-7 stable/train will continue to be tested in our
3rd party periodic pipeline and builds of CentOS-7
stable/train will be qualified here [2][3].

Also we are keeping c7 multinode containers job to
maintain c7 coverage.

[1]: https://review.opendev.org/#/c/738375
[2]: https://review.rdoproject.org/zuul/builds?pipeline=openstack-periodic-integration-stable2-centos7
[3]: http://lists.openstack.org/pipermail/openstack-discuss/2020-July/016157.html

Change-Id: I9f274c3866b214373c3562b0d6b480d15df1ce37
Signed-off-by: Bhagyashri Shewale <bshewale@redhat.com>