Donot trigger c7 container multinode job on stable/train
This job was kept initially to maintain coverage on c7, since
there is no user of CentOS-7 train, so we are removing this
job.

A periodic version of the job will be still running in the
upstream infra, https://review.opendev.org/#/c/752029/

Since https://review.opendev.org/#/c/752220/ already removes
the same job but it got triggered as it is defined in job
templates, so removing stable/train branch part from job
fixes it.

Added tripleo-ci-centos-8-content-provider to fix job
freeze issue.

Change-Id: I3ed7e99b7bc27094ab4347a4f883afc3787e422e
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>