pytest-html report for tcib container builds
Consolidate tcib build results in a html report running
pytest to check if the containers were built. For each
container image, a test is run to check if the image exists
and a link to the logs is attached to the report.

As part of the container build simplification and usability
improvements, the kolla and tripleo building methods have
been split into separate templates.

This change also removes the dry-runs in the build script that
were supposed to get the list of expected containers. Instead,
we replace the dry-runs with the tripleo_containers.yaml file.
This avoids confusion in counting expected containers vs the
actual built ones.

Change-Id: I1bb9353c2b5c79f55ab39d9cfcaa8f9617c28a34