Honour build_container_images and provider/consumer jobs
https://review.opendev.org/#/c/752685 changed all
jobs to use trunk.registry.rdoproject.org registry
for all rdo third party jobs but in RDO we run jobs
that build/deploy containers so we have to use
the local registry that is setup in jobs instead of relying on
external registeries.

This patch ensures correct registeries are used.

Change-Id: I9245d4c89e63ec71a102240e9ceaad18b5b14878