Switch scenario012 to non-voting
Scenario012 is overriding the bridge configuration in THT rather than
using what we configure in the standalone parameters. This leads to it
breaking when we try and fix the upgrade job. We need to adjust the
NeutronBridgeMappings to not be in the ci environment file but will
break the job while this is in flight.

Change-Id: I5c8aaba58171c55817dfb909cd18278a4a0305d5
Related-Bug: #1895822