Run only needed playbooks in undercloud jobs
In these jobs overcloud and upgrade playbooks also
runs which is not needed in undercloud jobs, so
let's set the playbooks var to run only the needed
playbooks.

Until https://review.opendev.org/#/c/759912/
merged it was not an issue as due to "tags" tasks
used to skip in the extra playbooks.

Depends-On: https://review.opendev.org/#/c/763519/
Change-Id: Ie1924646efd7333ce84b2a4a18c2cfc80edb7d51