toci- ensure the gating repo is release aware
The gating repo needs to be release aware.
This is REQUIRED for upgrade jobs where the release
changes during the course of the job.

This is also nice for transparency to ensure
the gating repo is set to the correct content-provider

related-review: https://review.opendev.org/c/763747
Please see the bug for more details

Partial-Bug: #1907122
Change-Id: I10d932e00f0639af8151f66f0751e48dae7b71ba