Allow to override extra_args & bm provisioning
We are adding extra_args & baremetal_provision to allowed_override to
make them overridable which enables us to test "provisioning via
metalsmith" feature in older branches and in downstream reusing fs001
and not maintain an entirely new featureset for these few settings.

baremetal_provision allows provisioning of node using metalsmith
workflow instead of heat/nova, This is true by default for ussuri+
branches.[1]

Also, we need to pass --disable-validations in deployment command when
we provision nodes via metalsmith which is set by default in
ussuri+[2]. This can be set using extra_args.

[1] https://opendev.org/openstack/tripleo-quickstart/src/branch/master/config/general_config/featureset001.yml#L231-L236
[2] https://opendev.org/openstack/tripleo-quickstart/src/branch/master/config/general_config/featureset001.yml#L72-L78

Change-Id: Ied37aa4ed8592142704446d20a3665345e46bddc