Removed unused db and wal
Ceph cluster is now provisioned by cephadm using a spec
where disks and logical volumes can be defined. However,
cephadm takes care about defining the disk layout if
nothing is specified.
The purpose of the patch is to remove the unused logical
volumes and keep the one needed to build the osd in CI.

Depends-On: I116a0e9d7fadca525c1e8f5328df77235379c09c
Change-Id: I26ae3b3a1ea9ce99d7855d6484bd42fef9360e76