Adding a check to see if os is stream os.
In centos 8 stream nodes some repos are not being updated to stream repos (they are still centos 8 repos) as --stream arg is not being set in tripleo-repos command. Also later in repo urls $stream is not rendered and container image build fails. So this patch:

1. Sets "is_stream_node"=false
2. If os is stream it overwrites "is_stream_node" to true
3. If "is_stream_node" == "True" it adds arg --stream to tripleo-repos command
4. Adds --volume flag to openstack tripleo container image build

I am adding a task to read /etc/os-release and get the name "Stream" as I could not find an ansible_distribution variable that shows it is a stream node https://termbin.com/fuke

Change-Id: I694404ea5bff3a21ca8437c6213bd0b0eadb572f