Remove standalone/undercloud from multinode layouts +deprecate old
As part of CI optimisation at [1] this deprecates the following
multinode templates:

  * tripleo-multinode-baremetal-minimal
  * tripleo-multinode-baremetal-full
  * tripleo-multinode-container-full
  * tripleo-multinode-container-minimal
  * tripleo-multinode-tripleoci

These are still used in some places such as [2][3] so we need
to remove those before we can delete the templates alltogether.

This also removes the standalone and undercloud jobs from the
multinode template. It is confusing to have them there and
increases the maintenance burden for the tripleo-ci team.

[1] https://review.opendev.org/q/topic:tripleo-ci-reduce
[2] https://review.opendev.org/c/openstack/os-apply-config/+/777097
[3] https://review.opendev.org/c/openstack/os-refresh-config/+/777098

Change-Id: Ie6473759383bd4b903cbe56fba9fee75a24154c4