Adds irrelevant-files to multinode-branchful content providers
As part of [1] this adds irrelevant-files to the content providers
victoria/ussuri/train that come with the multinode-branchful
template.

All of the jobs in this template share the same irrelevant-files
so if those don't run the content-providers should not either.
This is aiming to prevent situations as seen in [2] where change
to zuul.d/standalone.yaml gives completely unnecessary runs for
the victoria/ussuri/train providers (not used by any jobs there).

The depends-on here is not relevant to the patch but required to
get the victoria content-provider to pass.

[1] https://review.opendev.org/q/topic:tripleo-ci-reduce
[2] https://review.opendev.org/c/openstack/tripleo-ci/+/779447

Depends-On: https://review.opendev.org/c/openstack/tripleo-common/+/782367

Change-Id: I712c4011f77d6e95817c4653df001587e8ad7b23