Enable container-tools:3.0 before building images
We need to make sure proper modules are enabled on the host
before running the container-tools.

By default container-tools:rhel8 module is enabled on
the centos stream 8 nodepool image which brings higher version
of podman leading to break overcloud deployment.

We need to make sure it needs to be disabled and proper
container-tools modules needs to be enabled.

Related-Bug: #1922537

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I52244bcebf8e986f888e31b7ae51813f5dad1719