Use dib_dnf_module_streams for enabling modules
https://review.opendev.org/c/openstack/tripleo-ci/+/784887
enables container-tools during image build process but
it does not sets the proper modules so removing this
hack and switching to dib_dnf_module_streams var
to set it via DIB.

It also adds the same var for c7 with empty value for
not breaking the c7 image build flow.

It also adds -e in the build-images.sh.j2 script to catch
the errors early.

Depends-On: https://review.opendev.org/c/openstack/diskimage-builder/+/785138

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I769a0986c3258372ed8a421d7c2bf02e2bf8073e