Stop overriding manila tempest params
Manila's CephFS driver that we're testing
in scenario004 started supporting
snapshots and snapshot cloning, and our
tempest opts are incompatible.

Since tempestconf has a way to detect
the right options, lets stop overriding
in the job definitions.

Periodic RDO jobs will need to be fixed
in the same manner.

Depends-On: https://review.rdoproject.org/r/c/rdoinfo/+/33233
Related-Bug: #1922784
Change-Id: I4fbe340a110cb0b99e786844760af134672cda37