Add jobs for overcloud-hardened-uefi-full-centos-8
This change duplicates the
tripleo-buildimage-overcloud-hardened-full-centos-8* jobs to add
tripleo-buildimage-overcloud-hardened-uefi-full-centos-8* jobs for
master and wallaby.

The overcloud-hardened-full and overcloud-hardened-uefi-full images
are identical except that overcloud-hardened-uefi-full with boot on
both legacy BIOS and UEFI systems (overcloud-hardened-full will only
boot on legacy BIOS). As such there will be a follow-up change to
remove the tripleo-buildimage-overcloud-hardened-full-centos-8 from
running on master and wallaby.

Change-Id: I92170acf42b9c1850acc060653629cef15b116bf
Blueprint: whole-disk-default
Depends-On: Id5c8ba02762ab87fab5f8121dda37b463466c050