Handle overcloud-hardened-uefi-full in oooci-build-images
This change will generate a overcloud-hardened-uefi-full.qcow2.md5 if
build-images.sh builds a overcloud-hardened-uefi-full.qcow2.

Unlike overcloud-full and ironic-python-agent, no tar file is created
for overcloud-hardened-uefi-full.qcow2 because it is distributed as a
single qcow2 file.

This change also ensures
buildimage-overcloud-hardened-uefi-full-centos-8 only runs on master.
Gating is broken on stable/victoria and earlier because it is running
there also.

Change-Id: Ie2f7f5c5cc4a17d0e3514e5fc1accadcfa3b009a
Blueprint: whole-disk-default
Closes-Bug: #1933309