Install tripleo-operator-ansible from git for baremetal
In baremetal deployment, quickstart execution is done from
zuul node and quickstart depends on tripleo-operator-ansible.

Since https://review.opendev.org/c/openstack/tripleo-quickstart/+/790701
already removed the git installation of tripleo-operator-ansible
and we cannot install it via rpm so the failure is legit.

In order to fix that, we need to install tripleo-operator-ansible
from git on zuul node.

Related-Bug: #1938638

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I1fd588f02195dc09620f4950029faacb98b85eed