Move remove_tags value to skip_tag_list
https://review.opendev.org/c/openstack/tripleo-ci/+/758373 added
support for removing tags from the job config.

It is basically removing the tags from the whole tag list.
Instead of removing the tag, we can put it in the skip list.

It will help us from running the plays twice irrespective of
tags defined at role level or play level.

It also solves the issues of running build-test-packages on
consumer jobs where build tags in removed, is coming in
consolidation patch of build-test-packages [1].

[1]. https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/790926

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I072c682ce811871a7c06fb18b31d76f2e9397643