Adds tripleo-repos/-operator-ansible collection push jobs
In [1] a release job was added for tripleo-operator. Now we need
to add the same for the tripleo-repos collection. This proposes
we move the playbook, secret and job definition to here so that
both jobs can share them. Then the jobs can be wired up into their
respective repo zuul layout files.

Once we are happy with testing (this needs to merge so we can test)
we can merge [2] and remove the duplication in t-o-a.

[1] https://review.opendev.org/c/openstack/tripleo-operator-ansible/+/803588
[2] https://review.opendev.org/c/openstack/tripleo-operator-ansible/+/810259
Change-Id: I09b24c4782db2e7cad117fb98a70a1c3d36a4301