Ensure upgrade job read the proper hash
From time to time, due a promotion happening between the time
that the content provider job executes, and the upgrade job
the upgrade job fails because it can not find the hash in the
content provider. This is because the upgrade job, instead of
rely on the zuul job variable passed by the content provider,
it reads in execution time, the current-tripleo hash from
rdo registry. This patch ensure that if there is a zuul variable
passed by content provider, it will use that hash, instead of
read it from rdo registry.

Related-Bug: #1946659
Change-Id: I273243ced8b83aecec9da775e7f7a3c3865b4fc3