Allow override for install and target hash in emit releases undercloud
As described in related-bug, this allows us to override the install
and target delorean hash for use by consumer jobs. There was a previous
attempt with [1] which added the ability to override the target hash.

As commented in the bug (see comment #7) we need to override both target
and install versions. Since this is nested bash/jinja/python :/ we do this
by passing the content provider branches as a string
"branch1:hash1;branch2:hash2" generated in the jinja templating
which is then decoded on the emit-releases python side.

[1] https://review.opendev.org/c/openstack/tripleo-ci/+/813629
Co-Authored-By: Marios Andreou <marios@redhat.com>
Related-Bug: 1946659

Change-Id: I9e0162f88cf262957234bf946ad3c013f6213891