[workaround]Set permission for /dev/pts/ptmx to 666
When forking a pty as regular user, we get a failure to
read/write on /dev/ptmx since it's just a symlink to /dev/pts/ptmx.

It happens only on CS9 on all the nodes and gives error
OSError: out of pty devices.

Adding the workaround by setting the permission for /dev/pts/ptmx to
666 temprorly fixes the issue in CI.

Proper fix will come later.

Related-Bug: #1950176

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I1a0327b09f53beee1668a6ab4bb8380cc6b46446