Re-run repo-setup in image/container run plays after pre mirror setup
In RDO jobs the mirror-info-fork is included after the repo-setup
role. This means we don't have NODEPOOL_CENTOS_MIRROR available
in /etc/ci/mirror_info.sh and we use mirror.stream.centos.org
which is proving unreliable eg 404 for 9-stream/AppStream [1].

This patch re-runs the repo-setup role in the run plays for
containers and image builds. The mirror-info-fork is run in pre so
this way NODEPOOL_CENTOS_MIRROR is available when we do repo-setup
e.g. mirror used at [2]

[1] https://logserver.rdoproject.org/67/36267/36/check/periodic-tripleo-ci-build-containers-centos-9-push-master/575a6f3/logs/container-builds/c9967fc5-5f3b-4c06-b93b-eff2cd775c84/base/os/manila-base/manila-base-build.log
[2] https://logserver.rdoproject.org/17/36817/13/check/periodic-tripleo-centos-9-buildimage-overcloud-hardened-uefi-full-master/d4e06e9/yum.repos.d/quickstart-centos-appstreams.repo

Co-Authored-By: Douglas Viroel <viroel@gmail.com>
Change-Id: Ibb5617c84b7a31a320aebecc82d980ba8e9751b6