Add undercloud-ffu-wallaby job.
This patch adds all the missing bits to be able to perform
an Undercloud upgrade from Train to Wallaby in the CI. This
type of job is identified by the ffu_undercloud_upgrade keyword
and runs a similar set of steps as the Undercloud upgrade CI job,
relying on the tripleo-upgrade role to setup and run the upgrade
of the node.

The emit_releases_file script has been also modified to cover this
ffu jump from Train to Wallaby, as well as understanding the keyword
ffu_undercloud_upgrade (up to know only the ffu_overcloud_upgrade
existed).

This patch depends on the addition of a new featureset065 in which
the right parameters are being set.

Depends-On: https://review.opendev.org/c/openstack/tripleo-quickstart/+/822714

Change-Id: I876f7b2e03511b27cc1f02638964945c9518e503