Remove fwupd-redfish.conf file from overcloud images
Right now, the fwupd package is attempting to load the module ipmi_si
and this one is failing to be load. This patch customize the qcow2 image
removing the fwupd-redfish.conf that is causing this problem.
It is, right now hard to just exclude the package because of the
dependences, as well as downgrade, because of how the repositories are
configured in the image, so the best approach to temporarly solve the
problem is in this way with virt-customize.

Related-Bug: #1956441

Signed-off-by: Amol Kahat <amolkahat@gmail.com>

Change-Id: I07b2344f85c6997c10c3e50a86ba8e16dfcb1355