Allow to push the archives/images to another location
While it makes sense in some cases to get the final archive in the user
home directory, we might want to allow people to push those achives
somewhere else in the tree.

Using the new "oooci_image_build_archive_dest" parameter, we can
override the final location; it defaults to ansible_user_dir, in order
to not change the current behavior.

This patch also takes the opportunity to sanitize the "workspace" as a
source, especially in the "post.yaml" task file in order to ensure we're
using the same location, especially if we override the "workspace"
parameter.

Change-Id: Ie745184286f9b9db546d2853a53554ae7e11c23a