Centos8 master jobs teardown - c8 on train/ussuri/victoria/wallaby
This adjusts branches: match for the centos8 jobs so we avoid
running these for master. C8 jobs will only run on train ussuri
victoria and wallaby branches. The master branch will be c9 only
moving forward.

See [1] for related patches that depends-on this across the
tripleo repos. See [2] for related dummy nodeset required for
zuul 3rd party on this review.

[1] https://review.opendev.org/q/topic:c8_teardown_master
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/39045

Change-Id: If62f83de171de4e7651b58b70c31bea005710c71